animationメソッドをViewへ移動
